# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample authentication and authorization API built with Go
    and Gin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Authentication and Authorization API
  version: 1.1.0
paths:
...
  /auth/github/callback:
    get:
      description: Handles GitHub OAuth2 callback and returns JWT tokens
      parameters:
      - description: State token
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: GitHub OAuth2 Callback
      tags:
      - social
  /auth/github/login:
    get:
      description: Redirects user to GitHub OAuth2 login page
      parameters:
      - description: Frontend callback URL
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect
          schema:
            type: string
      summary: GitHub OAuth2 Login
      tags:
      - social
  /auth/google/callback:
    get:
      description: Handles Google OAuth2 callback and returns JWT tokens
      parameters:
      - description: State token
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth2 Callback
      tags:
      - social
  /auth/google/login:
    get:
      description: Redirects user to Google OAuth2 login page
      parameters:
      - description: Frontend callback URL
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Redirect
          schema:
            type: string
      summary: Google OAuth2 Login
      tags:
      - social
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" + your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"